Update 02-basics.md
Previous: A unique `userId` identifies a user. You should use this id whenever you *a* referring to a user. Access the id of a signed-in user through the `auth` field of the `Session` object.

Update: A unique `userId` identifies a user. You should use this id whenever you *are* referring to a user. Access the id of a signed-in user through the `auth` field of the `Session` object.